move all doe generators into single module, move tests to
