+++ openmdao/drivers/doe_generators.py
@@ -16,7 +16,75 @@ from openmdao.drivers.doe_driver import DOEGenerator
 import pyDOE
 
 
-class _Generator(DOEGenerator):
+class UniformGenerator(DOEGenerator):
+    """
+    DOE case generator implementing the Uniform method.
+
+    Attributes
+    ----------
+    _num_samples : int
+        The number of samples to run.
+    _seed : int or None
+        Random seed.
+    """
+
+    def __init__(self, num_samples=1, seed=None):
+        """
+        Initialize the UniformGenerator.
+
+        Parameters
+        ----------
+        num_samples : int, optional
+            The number of samples to run. Defaults to 1.
+
+        seed : int or None, optional
+            Seed for randon number generator.
+        """
+        super(UniformGenerator, self).__init__()
+
+        self._num_samples = num_samples
+        self._seed = seed
+
+    def __call__(self, design_vars):
+        """
+        Generate case.
+
+        Parameters
+        ----------
+        design_vars : dict
+            Dictionary of design variables for which to generate values.
+
+        Yields
+        ------
+        list
+            list of name, value tuples for the design variables.
+        """
+        if self._seed is not None:
+            np.random.seed(self._seed)
+
+        for i in range(self._num_samples):
+            sample = []
+
+            for (name, meta) in iteritems(design_vars):
+                values = []
+
+                for k in range(meta['size']):
+                    lower = meta['lower']
+                    if isinstance(lower, np.ndarray):
+                        lower = lower[k]
+
+                    upper = meta['upper']
+                    if isinstance(upper, np.ndarray):
+                        upper = upper[k]
+
+                    values.append(np.random.uniform(lower, upper))
+
+                sample.append((name, np.array(values)))
+
+            yield sample
+
+
+class _pyDOE_Generator(DOEGenerator):
     """
     Base class for DOE case generators implementing methods from pyDOE.
 
@@ -39,7 +107,7 @@ class _Generator(DOEGenerator):
             The number of evenly spaced levels between each design variable
             lower and upper bound. Defaults to 2.
         """
-        super(_Generator, self).__init__()
+        super(_pyDOE_Generator, self).__init__()
         self._levels = levels
 
     def __call__(self, design_vars):
@@ -117,7 +185,7 @@ class _Generator(DOEGenerator):
         pass
 
 
-class FullFactorialGenerator(_Generator):
+class FullFactorialGenerator(_pyDOE_Generator):
     """
     DOE case generator implementing the Full Factorial method.
     """
@@ -139,7 +207,7 @@ class FullFactorialGenerator(_Generator):
         return pyDOE.fullfact([self._levels] * size)
 
 
-class PlackettBurmanGenerator(_Generator):
+class PlackettBurmanGenerator(_pyDOE_Generator):
     """
     DOE case generator implementing the Plackett-Burman method.
     """
@@ -171,7 +239,7 @@ class PlackettBurmanGenerator(_Generator):
         return doe
 
 
-class BoxBehnkenGenerator(_Generator):
+class BoxBehnkenGenerator(_pyDOE_Generator):
     """
     DOE case generator implementing the Box-Behnken method.
